move top-level hokusai modules into lib and services
