+++ MANIFEST.in
@@ -1,1 +1,3 @@
 include hokusai/lib/*
+include hokusai/services/*
+include hokusai/commands/*
