+++ hokusai/__init__.py
@@ -1,2 +1,2 @@
-from hokusai import representers
+from hokusai.lib import representers
 from hokusai.commands import *
