+++ hokusai/commands/deploy.py
@@ -1,6 +1,6 @@
-from hokusai.command import command
-from hokusai.common import print_green
-from hokusai.deployment import Deployment
+from hokusai.lib.command import command
+from hokusai.lib.common import print_green
+from hokusai.services.deployment import Deployment
 
 @command
 def deploy(context, tag):
