+++ hokusai/lib/common.py
