+++ hokusai/services/ecr.py
@@ -3,7 +3,7 @@ import base64
 import boto3
 from botocore.exceptions import BotoCoreError
 
-from hokusai.config import config
+from hokusai.lib.config import config
 
 class ECR(object):
   def __init__(self):
