+++ test/unit/test_command.py
@@ -3,7 +3,7 @@ from mock import patch
 from test import HokusaiUnitTestCase
 from test.utils import captured_output
 
-from hokusai.command import command
+from hokusai.lib.command import command
 
 @command
 def foo_command(foo):
@@ -13,12 +13,12 @@ def foo_command(foo):
     raise ValueError('Bad command')
 
 class TestCommand(HokusaiUnitTestCase):
-  @patch('hokusai.command.sys.exit', retval=True)
+  @patch('hokusai.lib.command.sys.exit', retval=True)
   def test_command_exits_with_retval(self, mocked_sys_exit):
     foo_command('Ohai!')
     mocked_sys_exit.assert_called_once_with(0)
 
-  @patch('hokusai.command.sys.exit', retval=True)
+  @patch('hokusai.lib.command.sys.exit', retval=True)
   def test_command_catches_exceptions(self, mocked_sys_exit):
     with captured_output() as (out, err):
       foo_command(False)
